Refactor - put parsing tasks in own class
